realisations/serializers.py
from rest_framework import serializers
from .models import Realisation

class RealisationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Realisation
        fields = '__all__'