requirements.txt
# Django et composants principaux
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0,<3.16.0
djangorestframework-simplejwt>=5.3.0,<5.6.0
django-cors-headers>=4.3.0,<4.8.0
djoser>=2.2.0,<2.4.0
drf-yasg>=1.21.0,<1.22.0

# Base de données
psycopg2-binary>=2.9.9,<2.10.0  # Pour PostgreSQL
dj-database-url>=2.1.0,<2.2.0   # Pour la gestion des URL de connexion DB

# Sécurité
python-dotenv>=1.0.0,<1.2.0
PyJWT>=2.8.0,<2.10.0
cryptography>=41.0.0,<45.0.0

# Utilitaires
Pillow>=10.0.0,<10.1.0  # Pour le traitement d'images
pyyaml>=6.0.0,<6.1.0
uritemplate>=4.1.0,<4.2.0
pytz>=2023.3

# Tests
pytest>=7.4.0,<8.4.0
pytest-django>=4.5.0,<4.11.0
pytest-cov>=4.1.0,<4.2.0
coverage>=7.3.0,<7.4.0

# Outils de développement (optionnels)
black>=23.9.0,<24.0.0
isort>=5.12.0,<6.1.0
flake8>=6.1.0,<7.0.0
pylint>=3.0.0,<3.4.0
bandit>=1.7.5,<1.8.0

# Documentation OpenAPI
# Added drf-spectacular for schema generation
drf-spectacular>=0.26.0,<0.27.0
